+++ async_extensions/processes.py
@@ -1,3 +1,4 @@
 from anyio import open_process, run_process
+from anyio.abc import Process
 
-__all__ = ("open_process", "run_process")
+__all__ = ("Process", "open_process", "run_process")