+++ async_extensions/types.py
@@ -1,10 +1,8 @@
 from solus import Singleton
 
-__all__ = ("NoDefault", "no_default")
-
 
 class NoDefault(Singleton):
-    """Represents the absence of default values."""
+    pass
 
 
 no_default = NoDefault()