+++ CHANGELOG.md
@@ -2,6 +2,27 @@
 
 <!-- changelogging: start -->
 
+## 3.0.0 (2024-03-18)
+
+### Features
+
+- Added `as_completed`, `as_completed_results` and `as_completed_tagged_results`.
+- Added `NormalResult[T] = Result[T, NormalError]`.
+- Added `normal_result_of` to catch errors and wrap them.
+- Added `TaggedResult[T, E]` for results tagged with `int`.
+- Added `NormalTaggedResult[T] = TaggedResult[T, NormalError]`.
+- Added `normal_tagged_result_of`, akin to `normal_result_of` but attaching some tag.
+- Added `async_iter_unchecked` and `async_next_unchecked` which do not check whether the
+  argument has the appropriate type; also added `async_iter_any_iter` to convert `AnyIterable[T]`
+  into `AsyncIterator[T]`.
+- Exported `DEFAULT_BACKEND` and `DEFAULT_SHIELD`.
+- Exported `Process` and `async_close`.
+
+### Changes
+
+- Renamed `MemorySendChannel` and `MemoryReceiveChannel` to
+  `MemorySendStream` and `MemoryReceiveStream` respectively.
+
 ## 2.0.2 (2024-02-26)
 
 No significant changes.